Move records and rename to `/state/old`.
