+++ applications/desktop/src/notebook/store.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { createStore, applyMiddleware, combineReducers } from "redux";
 
-import type { AppState } from "@nteract/core/src/records";
+import type { AppState } from "@nteract/core/src/state";
 
 import middlewares from "./middlewares";
