+++ packages/core/src/epics/websocket-kernel.js
@@ -14,8 +14,6 @@ import { of } from "rxjs/observable/of";
 import { from } from "rxjs/observable/from";
 import { merge } from "rxjs/observable/merge";
 
-import type { AppState } from "../records";
-
 import { kernels, shutdown } from "rx-jupyter";
 import { v4 as uuid } from "uuid";
