+++ packages/core/src/index.js
@@ -6,7 +6,7 @@ import * as reducers from "./reducers";
 import * as themes from "./themes";
 import * as epics from "./epics";
 import * as selectors from "./selectors";
-import * as records from "./records";
+import * as records from "./state/old";
 import * as state from "./state";
 
 export * from "./components";
